+++ src/components/DropArrow/DropArrow.module.css
@@ -1,98 +1,97 @@
 .drop_container {
-    display: flex;
-    flex-wrap: nowrap;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 1rem;
-    flex-direction: column;
-    color: #ffffff;
-    font-size: 1.5rem;
-    cursor: pointer;
-    position: relative;
+  display: flex;
+  flex-wrap: nowrap;
+  justify-content: flex-end;
+  align-items: center;
+  gap: 1rem;
+  flex-direction: column;
+  color: black;
+  font-size: 1.5rem;
+  cursor: pointer;
+  position: relative;
 }
 
 .drop_title {
-    display: flex;
-    flex-wrap: nowrap;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 1rem;
-    flex-direction: row;
-    color: #ffffff;
+  display: flex;
+  flex-wrap: nowrap;
+  justify-content: flex-end;
+  align-items: center;
+  gap: 1rem;
+  flex-direction: row;
+  color: black;
 }
 
 .drop_circle {
-    width: 3rem;
-    height: 3rem;
-    border-radius: 50%;
-    background-color: rgb(105, 105, 105);
+  width: 3rem;
+  height: 3rem;
+  border-radius: 50%;
+  background-color: rgb(105, 105, 105);
 }
 
 .drop_arrow {
-    font-family: arial;
-    font-weight: 900;
+  font-family: arial;
+  font-weight: 900;
 }
 
 @keyframes dropping-down {
-    from {
-        height: 0px;
-    }
+  from {
+    height: 0px;
+  }
 
-    to {
-        height: 143px;
-    }
+  to {
+    height: 143px;
+  }
 }
 
 @keyframes dropping-up {
-    from {
-        height: 143px;
-    }
+  from {
+    height: 143px;
+  }
 
-    to {
-        height: 0px;
-    }
+  to {
+    height: 0px;
+  }
 }
 
 .drop_list {
-    position: absolute;
-    top: 47px;
-    left: 50px;
-    display: flex;
-    gap: 25px;
-    flex-direction: column-reverse;
-    z-index: 1;
-    text-align: right;
-    overflow: hidden;
-    width: 130px;
+  position: absolute;
+  top: 47px;
+  left: 50px;
+  display: flex;
+  gap: 25px;
+  flex-direction: column-reverse;
+  z-index: 1;
+  text-align: right;
+  overflow: hidden;
+  width: 130px;
 }
 
 .drop_list_open {
-    animation-name: dropping-down;
-    animation-duration: 0.5s;
-    height: 143px;
+  animation-name: dropping-down;
+  animation-duration: 0.5s;
+  height: 143px;
 }
 
 .drop_list_close {
-    animation-name: dropping-up;
-    animation-duration: 0.5s;
-    height: 0px;
+  animation-name: dropping-up;
+  animation-duration: 0.5s;
+  height: 0px;
 }
 
-
 .link_main {
-    &:hover {
-        opacity: 50%;
-    }
+  &:hover {
+    opacity: 50%;
+  }
 }
 
 .link_profile {
-    &:hover {
-        opacity: 50%;
-    }
+  &:hover {
+    opacity: 50%;
+  }
 }
 
 .link_exit {
-    &:hover {
-        opacity: 50%;
-    }
+  &:hover {
+    opacity: 50%;
+  }
 }
