+++ src/pages/Workout/Workout.jsx
@@ -1,37 +1,20 @@
-import { useSelector, useDispatch } from "react-redux";
-import {
-  submitProgress,
-  backToInitial,
-} from "../../store/reducers/mainReducers";
+import { useSelector } from "react-redux";
+import DropArrow from "../../components/DropArrow/DropArrow";
+import WorkoutExercises from "../../components/WorkoutExercises";
 import MyProgress from "../../components/Workout progress/myProgress";
 import S from "./Workout.module.css";
 
 function Workout() {
-  const dispatch = useDispatch();
   const completeProgressSwitcher = useSelector(
     (state) => state.rootReducer.mainState.initialState,
   );
 
-  const writeMyProgress = () => {
-    if (completeProgressSwitcher === false) {
-      dispatch(submitProgress());
-    } else {
-      dispatch(backToInitial());
-    }
-  };
-
   return (
     <>
-      {completeProgressSwitcher === false ? " " : <div className={S.cover} />}
       <header>
         <div className={S.header__wrapper}>
-          <div className={S.header__logo}>
-            <img src="logo.png" alt="logo_image" className={S.logo} />
-          </div>
-          <div className={S.header__name}>
-            <div className={S.name__img} />
-            <p className={S.name__text}>Сергей</p>
-          </div>
+          <div className={S.header__logo} />
+          <DropArrow />
         </div>
       </header>
       <main className={S.workout__auth}>
@@ -47,35 +30,13 @@ function Workout() {
             title="video"
             frameBorder="0"
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-            allowfullscreen
+            allowFullScreen
           />
         </div>
+        {completeProgressSwitcher === false ? " " : <div className={S.cover} />}
         {completeProgressSwitcher === false ? " " : <MyProgress />}
         <section className={S["workout-authorized__info"]}>
-          <div className={S["workout-authorized__exercises"]}>
-            <h3 className={S["workout-authorized__exercises-title"]}>
-              Упражнения
-            </h3>
-            <ul className={S["workout-authorized__list"]}>
-              <li className={S.exercises__item}>
-                Наклон вперед (10 повторений)
-              </li>
-              <li className={S.exercises__item}>
-                Наклон назад (10 повторений)
-              </li>
-              <li className={S.exercises__item}>
-                Поднятие ног, согнутых в коленях <br />
-                (5 повторений)
-              </li>
-            </ul>
-            <button
-              onClick={writeMyProgress}
-              type="button"
-              className={S["workout-authorized__check"]}
-            >
-              Заполнить свой прогресс
-            </button>
-          </div>
+          <WorkoutExercises />
           <div className={S["workout-authorized__progress"]}>
             <div className={S.progress__wrapper}>
               <h3 className={S.progress__title}>
