+++ src/store/reducers/mainReducers.js
@@ -27,6 +27,7 @@ const initialState = {
     description:
       "Йога - это философия здорового образа жизни. Тот, кто занимается йогой, становится здоровее и выносливее, после занятий чувствует прилив сил, а также с новой силой может ощутить вкус к жизни. Благодаря комплексному воздействию упражнений происходит проработка всех групп мышц, тренировка суставов, улучшается циркуляция крови. Кроме того, упражнения дарят отличное настроение, заряжают бодростью и помогают противостоять стрессам.",
   },
+  
 };
 
 const mainReducers = createSlice({
